ai_manager/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.5
langchain==0.0.350
langchain-openai==0.0.2
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
# Бесплатные AI провайдеры
aiohttp==3.9.1
transformers==4.36.2
torch==2.1.2
# Для развертывания на сервере
gunicorn==21.2.0
docker==6.1.3
